+++ docs/changelog.md
@@ -2,3 +2,4 @@
 
 #### v0.0.2
 - Added version checks for software release
+- Added auth check for username and password validation